+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add new `Capybara/RSpec/PredicateMatcher` cop. ([@ydah])
 - Fix a false positive for `Capybara/SpecificFinders` when `find` with kind option. ([@ydah])
 - Add new `Capybara/RSpec/HaveSelector` cop. ([@ydah])
+- Add new `Capybara/ClickLinkOrButtonStyle` cop. ([@ydah])
 
 ## 2.18.0 (2023-04-21)